feat: enhance cabin management with create, edit, and delete functionalities; integrate settings update features

+++ src/services/apiCabins.js
@@ -1,3 +1,4 @@
+import { ca } from "date-fns/locale";
 import supabase, { supabaseUrl } from "./supabase";
 
 export async function getCabins(){
@@ -13,49 +14,46 @@ return data
 
 }
 
-export async function createCabin(cabinData) {
-
+export async function createUpdateCabin(cabinData,id) {
+  const hasImagePath=cabinData.image?.startsWith?.(supabaseUrl)
   const imageName=`${Math.random()}-${cabinData.image.name}`.replaceAll("/","");
-  const imagePath=`${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
-//1 create cabin
-const { data, error } = await supabase
-  .from('cabins')
-  .insert([{...cabinData,image:imagePath}])
-  .select()
 
+  const imagePath=hasImagePath?cabinData.image:`${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
+
+  console.log("imagePath",imagePath,hasImagePath,cabinData,supabaseUrl)
+let query=supabase.from('cabins')
+
+//1 create cabin
+  if(!id)
+ query= query.insert([{...cabinData,image:imagePath}])
+
+if(id)
+query= query.update({...cabinData,image:imagePath})
+  .eq('id', id)
+
+
+const {data,error}=await query.select()
 if(error){
 console.error(error);
 throw new Error('Failed to fetch cabins');
 }
 
-//2 Upload file using standard upload
+if(!hasImagePath){
+  //2 Upload file using standard upload
   const {  error:storageError } = await supabase.storage.from('cabin-images').upload(imageName, cabinData.image)
   if (storageError) {
-   await supabase
-  .from('cabins')
-  .delete()
-  .eq('id', data[0].id);
-  console.error(storageError);
-  throw new Error("cabin image could not be uploaded")
+    await supabase
+    .from('cabins')
+    .delete()
+    .eq('id', data[0].id);
+    console.error(storageError);
+    throw new Error("cabin image could not be uploaded")
   }
-return data
-
-}
-export async function updateCabin(cabinId, cabinData) {
-
-const { data, error } = await supabase
-  .from('cabins')
-  .update({ other_column: 'otherValue' })
-  .eq('some_column', 'someValue')
-  .select()
-
-  if(error){
-console.error(error);
-throw new Error('Failed to fetch cabins');
 }
 return data
 
 }
+
 
 
 export async function deleteCabin(cabinId) {
